Add files via upload
Added menu scene, variable clarifications, sound effects, and comments in the code. Everything should be more legible now, this is still a testing ground for the dice roll and come out roll mechanics which are some of the more finicky aspects of this game. As it stands they are functioning and adding onto those mechanics is the work of the final game.